Create search bar complete
I plan to wire the search bar function to the submission of the form very soon, but I learned a lot in this section. I didn't know that encodeURIComponent existed, but I ended up learning about it from external sources and it blew my mind. I am ecstatic at the opportunities this syntax will allow for